Adding Dockerfile to bypass "simple" building of dotnet project
The Railway builder for .net 8 is almost useless. It was using a version
of the SDK that was more than a year old, it was actually from one of
the .net 8 previews. That SDK would not even compile this blazor app.

Between that and the way it required the layout of the repo, it is
easier to just use a Dockerfile. Though it would be nice to not have
to maintain one.